# openapi/openapi.yaml
components:
  schemas:
    handlers.AgentEnrollRequest:
      properties:
        agent_docker_endpoint:
          description: AgentDockerEndpoint 是 manager 访问该 agent Docker 代理的地址。
          type: string
        agent_file_endpoint:
          description: AgentFileEndpoint 是 manager 访问该 agent 文件代理的地址。
          type: string
        agent_id:
          description: AgentID 是 agent 自报的稳定外部 ID，用于幂等 enroll 与节点复用。
          type: string
        agent_tls_ca_cert:
          description: AgentTLSCACert 是 agent 代理 TLS CA 证书内容，可为空。
          type: string
        agent_version:
          description: AgentVersion 是 agent 当前版本，用于节点巡检与升级判断。
          type: string
        max_apps:
          description: MaxApps 是节点可承载应用数量上限；nil 表示不限。
          type: integer
        metadata:
          additionalProperties: {}
          description: Metadata 是 agent 附加元数据，原样保存为 JSON。
          type: object
        name:
          description: Name 是节点展示名，空值由 service 使用 AgentID 兜底。
          type: string
        node_data_root:
          description: NodeDataRoot 是 agent 侧应用数据根目录。
          type: string
        node_resource:
          $ref: '#/components/schemas/handlers.AgentNodeResourceRequest'
        resource_snapshot:
          additionalProperties: {}
          description: ResourceSnapshot 是 agent enroll 时上报的资源快照，原样保存为 JSON。
          type: object
        sampled_at:
          description: SampledAt 是资源采样时间；为空时 handler 使用当前 UTC 时间兼容旧 agent。
          type: string
      required:
      - agent_id
      type: object
    handlers.AgentHeartbeatRequest:
      properties:
        agent_token:
          description: AgentToken 是 enroll 后签发的节点令牌，用于认证心跳来源。
          type: string
        agent_version:
          description: AgentVersion 是心跳时的 agent 版本。
          type: string
        metadata:
          additionalProperties: {}
          description: Metadata 是心跳时的附加元数据，覆盖节点当前元数据。
          type: object
        node_resource:
          $ref: '#/components/schemas/handlers.AgentNodeResourceRequest'
        resource_snapshot:
          additionalProperties: {}
          description: ResourceSnapshot 是心跳时的资源快照，覆盖节点当前快照。
          type: object
        sampled_at:
          description: SampledAt 是资源采样时间；为空时 handler 使用当前 UTC 时间兼容旧 agent。
          type: string
      required:
      - agent_token
      type: object
    handlers.AgentNodeResourceRequest:
      description: NodeResource 是 agent 心跳时可选的节点资源采样。
      properties:
        cpu_percent:
          description: CPUPercent 是节点 CPU 使用百分比；nil 表示 agent 未采集该指标。
          type: number
        disk_total_bytes:
          description: DiskTotalBytes 是节点磁盘总字节数。
          type: integer
        disk_used_bytes:
          description: DiskUsedBytes 是节点磁盘已用字节数。
          type: integer
        instance_count:
          description: InstanceCount 是采样时节点承载的实例数量。
          type: integer
        last_error:
          description: LastError 是 agent 侧采样失败原因；空字符串表示本次采样未报告错误。
          type: string
        memory_total_bytes:
          description: MemoryTotalBytes 是节点内存总字节数。
          type: integer
        memory_used_bytes:
          description: MemoryUsedBytes 是节点内存已用字节数。
          type: integer
        network_rx_bytes:
          description: NetworkRxBytes 是节点网络累计接收字节数。
          type: integer
        network_tx_bytes:
          description: NetworkTxBytes 是节点网络累计发送字节数。
          type: integer
      type: object
...
    handlers.ErrorResponse:
      properties:
        code:
          description: Code 是稳定的接口契约标识，一旦发布只增不改。
          example: APP_NOT_FOUND
          type: string
        message:
          description: Message 是面向前端展示的安全错误文案。
          example: 应用不存在
          type: string
      type: object
...
    service.AgentEnrollResult:
      properties:
        agent_token:
          description: AgentToken 是新签发的 agent 明文 token，仅在 enroll 响应中返回一次。
          type: string
        heartbeat_interval_seconds:
          description: HeartbeatIntervalSeconds 是 agent 后续上报心跳的建议间隔。
          type: integer
        node_id:
          description: NodeID 是 manager 为该 agent 分配或复用的 runtime node ID。
          type: string
      type: object
...
    service.NodeCurrentResourceResult:
      description: CurrentResource 是节点最近一次资源采样摘要；列表页用于展示当前资源状态。
      properties:
        cpu_percent:
          description: CPUPercent 是节点 CPU 使用百分比；nil 表示最近采样缺少该指标。
          type: number
        disk_total_bytes:
          description: DiskTotalBytes 是节点磁盘总字节数。
          type: integer
        disk_used_bytes:
          description: DiskUsedBytes 是节点磁盘已用字节数。
          type: integer
        instance_count:
          description: InstanceCount 是采样时节点承载的实例数量。
          type: integer
        last_error:
          description: LastError 是最近一次节点资源采样错误，空字符串表示无错误或未上报。
          type: string
        memory_total_bytes:
          description: MemoryTotalBytes 是节点内存总字节数。
          type: integer
        memory_used_bytes:
          description: MemoryUsedBytes 是节点内存已用字节数。
          type: integer
        sampled_at:
          description: SampledAt 是最近一次采样时间，统一输出 UTC RFC3339。
          type: string
      type: object
...
    service.RuntimeNodeResult:
      properties:
        agent_docker_endpoint:
          description: AgentDockerEndpoint 是 manager 访问该节点 Docker 代理的 HTTPS 地址。
          type: string
        agent_file_endpoint:
          description: AgentFileEndpoint 是 manager 访问该节点文件代理的 HTTPS 地址。
          type: string
        agent_id:
          description: AgentID 是 agent 自报的外部稳定 ID，用于幂等 enroll。
          type: string
        agent_version:
          description: AgentVersion 是最近一次 enroll 或 heartbeat 上报的 agent 版本。
          type: string
        current_resource:
          $ref: '#/components/schemas/service.NodeCurrentResourceResult'
        has_agent_token:
          description: HasAgentToken 表示节点是否已保存 agent token hash；不会返回 token 明文。
          type: boolean
        heartbeat_interval_seconds:
          description: HeartbeatIntervalSeconds 是 manager 告知 agent 的心跳间隔秒数。
          type: integer
        id:
          description: ID 是 manager 侧 runtime_nodes 主键。
          type: string
        last_probe_attempted_at:
          description: LastProbeAttemptedAt 是最近一次探测开始时间，空值表示尚未探测。
          type: string
        last_probe_error:
          description: LastProbeError 是最近一次探测失败原因，已由探测流程写入安全错误文本。
          type: string
        last_probe_failed_at:
          description: LastProbeFailedAt 是最近一次探测失败时间。
          type: string
        last_probe_ok_at:
          description: LastProbeOKAt 是最近一次探测成功时间。
          type: string
        max_apps:
          description: MaxApps 是节点可承载应用上限；nil 表示不限制。
          type: integer
        name:
          description: Name 是管理员后台展示的节点名称。
          type: string
        node_data_root:
          description: NodeDataRoot 是 agent 侧应用数据根目录。
          type: string
        probe_failure_streak:
          description: ProbeFailureStreak 是连续探测失败次数，用于前端展示节点健康风险。
          type: integer
        probe_success_streak:
          description: ProbeSuccessStreak 是连续探测成功次数，用于判断节点恢复稳定性。
          type: integer
        status:
          description: Status 是节点管理状态，active 才允许 agent 心跳继续更新。
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: 登录后获得的 JWT access token，前缀 "Bearer "。
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: 基于 Hermes Agent runtime 的多组织管理后台 API。
  title: Agent Runtime Manager API
  version: "1.0"
openapi: 3.1.0
paths:
  /agent/enroll:
    post:
      description: runtime agent 使用共享 enrollment secret 自动注册或刷新节点信息，并换取长效 agent token
      parameters:
      - description: Bearer enrollment_secret
        in: header
        name: Authorization
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handlers.AgentEnrollRequest'
                description: 自动注册请求
                summary: body
        description: 自动注册请求
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/service.AgentEnrollResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      summary: Agent 自动注册
      tags:
      - agent
  /agent/heartbeat:
    post:
      description: runtime agent 定期上报心跳及资源快照；鉴权通过请求体中的 agent_token 字段完成
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handlers.AgentHeartbeatRequest'
                description: 心跳请求（含 agent_token）
                summary: body
        description: 心跳请求（含 agent_token）
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/service.RuntimeNodeResult'
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      summary: Agent 心跳
      tags:
      - agent
...
servers:
- url: /api/v1